pyproject.toml
[tool.poetry]
name = "diy-rag"
version = "0.1.0"
description = ""
authors = ["siesto1elemento <rohitojha9720@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openparse = "^0.7.0"
sentence-transformers = "^4.1.0"
faiss-cpu = "^1.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
